Refactor logging imports and function names
