+++ excel_analysis/utils/registro.py
